Add support contact functionality (#286)
The updates include the addition of a support contact button in the topbar of the application layout. This button is linked to a "mailto" address, which allows users to send an email to the support team directly. The 'Support' label was also added to the internationalization files for multiple languages.